Extract UI and Common code
