+++ Assets/GracesGames/2DTileMapLevelEditor/Prefabs/LevelEditor.prefab
@@ -51,33 +51,22 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: fe68dddc1943d4bfcbd0a134ce7a81b4, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  LevelEditorUiPrefab: {fileID: 1352682877195322, guid: 036231413d74d4f71829b18988be0136,
+    type: 2}
+  FileBrowserPrefab: {fileID: 1760643962491650, guid: 4599b5ba3543545db92bb18160cea721,
+    type: 2}
   Height: 10
   Width: 16
   Layers: 10
   Tiles:
-  - {fileID: 4000010041083728, guid: f4a7e24c932b14c6488bda2f47a7ff45, type: 2}
-  - {fileID: 4000010041083728, guid: 6b8e52f28732647acb63c0f7698e0f47, type: 2}
-  - {fileID: 4000010041083728, guid: 6d502604d1507447e91d7dd3ba82c063, type: 2}
-  - {fileID: 4000013102667656, guid: 316a8557d38a54a2989f1ea1171f2a18, type: 2}
-  - {fileID: 4000010041083728, guid: 46266abbd7d62496b97d38f5f3448689, type: 2}
-  - {fileID: 4000010041083728, guid: 04d2aa2b3235d4f96a9bc3453962b46e, type: 2}
-  - {fileID: 4000010041083728, guid: cf524a5442a29460095163d69c9ba515, type: 2}
-  - {fileID: 4000010041083728, guid: b0a97a8cde6154196b19c89b29c07541, type: 2}
-  - {fileID: 4000010041083728, guid: 9ddaf934523c14d3ea69f1ff31e74a74, type: 2}
-  - {fileID: 4000010041083728, guid: e30cc65ad371442cca0dddf9cc71b51c, type: 2}
-  - {fileID: 4000010041083728, guid: b03ff37209fbc4a5b838776400df6abb, type: 2}
-  - {fileID: 4000010041083728, guid: 6a94c817bb06d43ed86b967bed96d531, type: 2}
-  - {fileID: 4000010041083728, guid: b1703c8734925405e848b4976d8c7b73, type: 2}
-  ButtonPrefab: {fileID: 1000010740908782, guid: 993ac6289fda64f09b05ea526639da57,
-    type: 2}
-  ButtonSize: 95
-  SelectedTileSize: 100
-  ButtonImageScale: 0.8
-  NoSelectedTileImage: {fileID: 21300000, guid: b0aa86e0f4ec845b295e83ba21730cd6,
-    type: 3}
+  - {fileID: 4000013102667656, guid: b34e418e5797844cfae30860de728df3, type: 2}
+  - {fileID: 4000010041083728, guid: d573bda6707bc405ca3f0a02c446696b, type: 2}
+  - {fileID: 4000010041083728, guid: efbbd2ae9aa8f460e82905a8fb0974c5, type: 2}
+  - {fileID: 4000010041083728, guid: bedde3dd8dfb64627aa0343cfb9a455b, type: 2}
+  - {fileID: 4000010041083728, guid: 00890ac1e002b45e491cb768cdaef64a, type: 2}
+  - {fileID: 4000010041083728, guid: 7d38d188e97d848ff920cb4cd2e3e373, type: 2}
+  - {fileID: 4000010041083728, guid: f7fd9cda4e1e24e2a85bb97a61f8cc81, type: 2}
+  - {fileID: 4000011169756728, guid: c5bf4424a05514be597de92e852ff421, type: 2}
+  - {fileID: 4000010041083728, guid: 2cece3ef8eba449e8a841ef15c589ee7, type: 2}
+  - {fileID: 4000010041083728, guid: 537f2cee0def540c897bcced1b17b98b, type: 2}
   FileExtension: lvl
-  LevelEditorUiPrefab: {fileID: 1352682877195322, guid: 036231413d74d4f71829b18988be0136,
-    type: 2}
-  FillCursor: {fileID: 2800000, guid: 06eea2f700f324cb88e80e9f58abca3f, type: 3}
-  FileBrowserPrefab: {fileID: 1760643962491650, guid: 4599b5ba3543545db92bb18160cea721,
-    type: 2}
