+++ Assets/GracesGames/2DTileMapLevelEditor/Prefabs/UI.meta
@@ -1,0 +1,9 @@
+fileFormatVersion: 2
+guid: 576a2ab3f06954f9bba93c2fb4098ccc
+folderAsset: yes
+timeCreated: 1513161175
+licenseType: Free
+DefaultImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
