+++ Assets/GracesGames/2DTileMapLevelEditor/Scenes/PlatformerDemo.unity
@@ -120,6 +120,11 @@ Prefab:
       propertyPath: tiles.Array.size
       value: 13
       objectReference: {fileID: 0}
+    - target: {fileID: 114203227262763276, guid: 840ad8b7ad2c249179d3bc1e62ace2f2,
+        type: 2}
+      propertyPath: Tiles.Array.size
+      value: 13
+      objectReference: {fileID: 0}
     - target: {fileID: 4476745195887002, guid: 840ad8b7ad2c249179d3bc1e62ace2f2, type: 2}
       propertyPath: m_LocalPosition.x
       value: 0
@@ -350,6 +355,84 @@ Prefab:
       propertyPath: Height
       value: 10
       objectReference: {fileID: 0}
+    - target: {fileID: 114203227262763276, guid: 840ad8b7ad2c249179d3bc1e62ace2f2,
+        type: 2}
+      propertyPath: Tiles.Array.data[0]
+      value: 
+      objectReference: {fileID: 4000010041083728, guid: f4a7e24c932b14c6488bda2f47a7ff45,
+        type: 2}
+    - target: {fileID: 114203227262763276, guid: 840ad8b7ad2c249179d3bc1e62ace2f2,
+        type: 2}
+      propertyPath: Tiles.Array.data[1]
+      value: 
+      objectReference: {fileID: 4000010041083728, guid: 6b8e52f28732647acb63c0f7698e0f47,
+        type: 2}
+    - target: {fileID: 114203227262763276, guid: 840ad8b7ad2c249179d3bc1e62ace2f2,
+        type: 2}
+      propertyPath: Tiles.Array.data[2]
+      value: 
+      objectReference: {fileID: 4000010041083728, guid: 6d502604d1507447e91d7dd3ba82c063,
+        type: 2}
+    - target: {fileID: 114203227262763276, guid: 840ad8b7ad2c249179d3bc1e62ace2f2,
+        type: 2}
+      propertyPath: Tiles.Array.data[3]
+      value: 
+      objectReference: {fileID: 4000013102667656, guid: 316a8557d38a54a2989f1ea1171f2a18,
+        type: 2}
+    - target: {fileID: 114203227262763276, guid: 840ad8b7ad2c249179d3bc1e62ace2f2,
+        type: 2}
+      propertyPath: Tiles.Array.data[4]
+      value: 
+      objectReference: {fileID: 4000010041083728, guid: 46266abbd7d62496b97d38f5f3448689,
+        type: 2}
+    - target: {fileID: 114203227262763276, guid: 840ad8b7ad2c249179d3bc1e62ace2f2,
+        type: 2}
+      propertyPath: Tiles.Array.data[5]
+      value: 
+      objectReference: {fileID: 4000010041083728, guid: 04d2aa2b3235d4f96a9bc3453962b46e,
+        type: 2}
+    - target: {fileID: 114203227262763276, guid: 840ad8b7ad2c249179d3bc1e62ace2f2,
+        type: 2}
+      propertyPath: Tiles.Array.data[6]
+      value: 
+      objectReference: {fileID: 4000010041083728, guid: cf524a5442a29460095163d69c9ba515,
+        type: 2}
+    - target: {fileID: 114203227262763276, guid: 840ad8b7ad2c249179d3bc1e62ace2f2,
+        type: 2}
+      propertyPath: Tiles.Array.data[7]
+      value: 
+      objectReference: {fileID: 4000010041083728, guid: b0a97a8cde6154196b19c89b29c07541,
+        type: 2}
+    - target: {fileID: 114203227262763276, guid: 840ad8b7ad2c249179d3bc1e62ace2f2,
+        type: 2}
+      propertyPath: Tiles.Array.data[8]
+      value: 
+      objectReference: {fileID: 4000010041083728, guid: 9ddaf934523c14d3ea69f1ff31e74a74,
+        type: 2}
+    - target: {fileID: 114203227262763276, guid: 840ad8b7ad2c249179d3bc1e62ace2f2,
+        type: 2}
+      propertyPath: Tiles.Array.data[9]
+      value: 
+      objectReference: {fileID: 4000010041083728, guid: e30cc65ad371442cca0dddf9cc71b51c,
+        type: 2}
+    - target: {fileID: 114203227262763276, guid: 840ad8b7ad2c249179d3bc1e62ace2f2,
+        type: 2}
+      propertyPath: Tiles.Array.data[10]
+      value: 
+      objectReference: {fileID: 4000010041083728, guid: b03ff37209fbc4a5b838776400df6abb,
+        type: 2}
+    - target: {fileID: 114203227262763276, guid: 840ad8b7ad2c249179d3bc1e62ace2f2,
+        type: 2}
+      propertyPath: Tiles.Array.data[11]
+      value: 
+      objectReference: {fileID: 4000010041083728, guid: 6a94c817bb06d43ed86b967bed96d531,
+        type: 2}
+    - target: {fileID: 114203227262763276, guid: 840ad8b7ad2c249179d3bc1e62ace2f2,
+        type: 2}
+      propertyPath: Tiles.Array.data[12]
+      value: 
+      objectReference: {fileID: 4000010041083728, guid: b1703c8734925405e848b4976d8c7b73,
+        type: 2}
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 100100000, guid: 840ad8b7ad2c249179d3bc1e62ace2f2, type: 2}
   m_IsPrefabParent: 0
