+++ Assets/GracesGames/2DTileMapLevelEditor/Scenes/SokobanDemo.unity
@@ -347,6 +347,11 @@ Prefab:
       propertyPath: MapSize.x
       value: 0.45
       objectReference: {fileID: 0}
+    - target: {fileID: 114203227262763276, guid: 840ad8b7ad2c249179d3bc1e62ace2f2,
+        type: 2}
+      propertyPath: SaveType
+      value: 0
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 100100000, guid: 840ad8b7ad2c249179d3bc1e62ace2f2, type: 2}
   m_IsPrefabParent: 0
