+++ Assets/GracesGames/Common/Scripts/Utilities.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 75e06a35c8b334ebfa409567d1f02f99
+timeCreated: 1513152228
+licenseType: Free
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
